markPrice selector + some props removed
